Updated Code Again
So I wrote it so that it mostly works. The only issue I am seeing is I am having trouble saying there is a parsing error if you don't have anything for the 2nd expression at 
Assignment → Identifier[[Expression]] = Expression..

Most of it has been tested to work but my code is a mess, but I really don't want to work on this anymore for a while because Ive spent a long while doing this.